+++ CHANGELOG.rst
@@ -19,7 +19,6 @@ Next Release (TBD)
 * bugfix:Timestamp Input: Fix regression where timestamps without any timezone
   information were not being handled properly
   (`issue 982 <https://github.com/aws/aws-cli/issues/982>`__)
-<<<<<<< HEAD
 * bugfix:Signature Version 4: You can enable Signature Version 4 for Amazon S3
   commands by running ``aws configure set default.s3.signature_version s3v4``
   (`issue 1006 <https://github.com/aws/aws-cli/issues/1006>`__,
@@ -30,6 +29,9 @@ Next Release (TBD)
 * feature:Binary File Input: Add support for reading file contents as binary
   by prepending the filename with ``fileb://``
   (`issue 1010 <https://github.com/aws/aws-cli/pull/1010>`__)
+* bugfix:Streaming Output File: Fix issue when streaming a response to a file
+  and an error response is returned
+  (`issue 1012 <https://github.com/aws/aws-cli/pull/1012>`__)
 
 
 1.6.2